burrego: big refactoring
Lots of code cleanup, better usage of wasmtime.

Also, rework codebase layout
