+++ crates/burrego/.gitignore
@@ -1,0 +1,3 @@
+target
+*.wasm
+*.tar.gz
