+++ crates/burrego/src/builtins/encoding.rs
