+++ crates/burrego/src/builtins/mod.rs
@@ -1,6 +1,7 @@
 use anyhow::Result;
 use std::collections::HashMap;
 
+pub(crate) mod builtins_helper;
 mod debugging;
 mod encoding;
 mod glob;
@@ -10,7 +11,9 @@ mod semver;
 mod strings;
 mod time;
 
-pub type BuiltinFunctionsMap =
+pub(crate) use builtins_helper::BUILTINS_HELPER;
+
+pub(crate) type BuiltinFunctionsMap =
     HashMap<&'static str, fn(&[serde_json::Value]) -> Result<serde_json::Value>>;
 
 pub fn get_builtins() -> BuiltinFunctionsMap {
