+++ crates/burrego/src/host_callbacks.rs
@@ -1,13 +1,7 @@
-use lazy_static::lazy_static;
-
-lazy_static! {
-    pub static ref DEFAULT_HOST_CALLBACKS: HostCallbacks = HostCallbacks::default();
-}
-
 /// HostCallback is a type that references a pointer to a function
 /// that can be stored and then invoked by burrego when the Open
 /// Policy Agent Wasm target invokes certain Wasm imports.
-pub type HostCallback = Box<dyn Fn(String) + Send + Sync + 'static>;
+pub type HostCallback = fn(&str);
 
 /// HostCallbacks defines a set of pluggable host implementations of
 /// OPA documented imports:
@@ -15,4 +9,21 @@ pub type HostCallback = Box<dyn Fn(String) + Send + Sync + 'static>;
 pub struct HostCallbacks {
     pub opa_abort: HostCallback,
     pub opa_println: HostCallback,
+}
+
+impl Default for HostCallbacks {
+    fn default() -> HostCallbacks {
+        HostCallbacks {
+            opa_abort: default_opa_abort,
+            opa_println: default_opa_println,
+        }
+    }
+}
+
+fn default_opa_abort(msg: &str) {
+    eprintln!("OPA abort with message: {:?}", msg);
+}
+
+fn default_opa_println(msg: &str) {
+    println!("Message coming from the policy: {:?}", msg);
 }
