+++ crates/burrego/src/lib.rs
@@ -1,1 +1,10 @@
-pub mod opa;
+mod builtins;
+mod evaluator;
+pub mod host_callbacks;
+mod opa_host_functions;
+mod policy;
+mod stack_helper;
+
+pub use builtins::get_builtins;
+pub use evaluator::Evaluator;
+pub use host_callbacks::HostCallbacks;
