+++ crates/burrego/test_data/trace/policy.rego
@@ -1,0 +1,9 @@
+package policy
+
+default main = false
+
+main {
+  trace(sprintf("input.message has been set to '%v'", [input.message]));
+  m := input.message;
+  m == "world"
+}
